requirements.txt
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.7
django-filter==1.0.4
django-letsencrypt==2.0.0
djangorestframework==3.6.3
gunicorn==19.7.1
Markdown==2.6.9
psycopg2==2.7.3.1
pytz==2017.2
six==1.10.0
static3==0.7.0
whitenoise==3.3.0
